bacphlip-service/pyproject.toml
[project]
name = "bacphlip-service"
version = "0.1.0"
description = "Bacphlip phage lifestyle prediction service"
readme = "README.md"
requires-python = ">=3.9,<3.10"
dependencies = [
    "bacphlip @ git+https://github.com/adamhockenberry/bacphlip.git",
    "biopython>=1.7",
    "pandas>=1.3,<2.0",
    "joblib>=1.0",
    "scikit-learn==0.24.2",
    "numpy>=1.19,<1.24",
    "loguru>=0.7",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "fastapi>=0.100",
    "uvicorn>=0.22",
    "setuptools",
]

[tool.uv]
override-dependencies = [
    "scikit-learn==0.24.2",
    "numpy>=1.19,<1.24",
]
